requirements.txt
Package                   Version
------------------------- --------
asgiref                   3.7.2
attrs                     23.1.0
Django                    4.2.4
django-filter             23.2
django-split-settings     1.2.0
djangorestframework       3.14.0
drf-spectacular           0.26.4
inflection                0.5.1
jsonschema                4.19.0
jsonschema-specifications 2023.7.1
pip                       22.3.1
psycopg2                  2.9.7
pytz                      2023.3
PyYAML                    6.0.1
referencing               0.30.2
rpds-py                   0.9.2
setuptools                65.5.1
sqlparse                  0.4.4
typing_extensions         4.7.1
uritemplate               4.1.1
wheel                     0.38.4
